refactor: 이벤트 payload 위치 변경
